src/test/java/Day_10/SmallALUTest.java
package Day_10;

import Year_2022.Day_10.SmallALU;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class SmallALUTest {

    private SmallALU smallALU;
    @BeforeEach
    public void setup(){
        smallALU = new SmallALU();
    }

    @Test
    public void testNoopOnce(){
        smallALU.operation("noop",0);
        assertEquals(1,smallALU.getCurrentX());
    }

    @Test
    public void testNoopMany(){
        smallALU.operation("noop",0);
        smallALU.operation("noop",0);
        smallALU.operation("noop",0);
        smallALU.operation("noop",0);
        smallALU.operation("noop",0);
        smallALU.operation("noop",0);
        assertEquals(1,smallALU.getCurrentX());
    }

    @Test
    public void testAddxPositive(){
        smallALU.operation("addx",10);
        assertEquals(11,smallALU.getCurrentX());
    }

    @Test
    public void testXByCycleNumber(){
        smallALU.operation("noop",0);
        smallALU.operation("addx",3);
        smallALU.operation("addx",-5);
        assertEquals(4,smallALU.getXByCycleNumber(4));
        assertEquals(-1,smallALU.getXByCycleNumber(5));
    }

    @Test
    public void testExample(){
        String input = "" +
                "addx 15\n" +
                "addx -11\n" +
                "addx 6\n" +
                "addx -3\n" +
                "addx 5\n" +
                "addx -1\n" +
                "addx -8\n" +
                "addx 13\n" +
                "addx 4\n" +
                "noop\n" +
                "addx -1\n" +
                "addx 5\n" +
                "addx -1\n" +
                "addx 5\n" +
                "addx -1\n" +
                "addx 5\n" +
                "addx -1\n" +
                "addx 5\n" +
                "addx -1\n" +
                "addx -35\n" +
                "addx 1\n" +
                "addx 24\n" +
                "addx -19\n" +
                "addx 1\n" +
                "addx 16\n" +
                "addx -11\n" +
                "noop\n" +
                "noop\n" +
                "addx 21\n" +
                "addx -15\n" +
                "noop\n" +
                "noop\n" +
                "addx -3\n" +
                "addx 9\n" +
                "addx 1\n" +
                "addx -3\n" +
                "addx 8\n" +
                "addx 1\n" +
                "addx 5\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "addx -36\n" +
                "noop\n" +
                "addx 1\n" +
                "addx 7\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "addx 2\n" +
                "addx 6\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "addx 1\n" +
                "noop\n" +
                "noop\n" +
                "addx 7\n" +
                "addx 1\n" +
                "noop\n" +
                "addx -13\n" +
                "addx 13\n" +
                "addx 7\n" +
                "noop\n" +
                "addx 1\n" +
                "addx -33\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "addx 2\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "addx 8\n" +
                "noop\n" +
                "addx -1\n" +
                "addx 2\n" +
                "addx 1\n" +
                "noop\n" +
                "addx 17\n" +
                "addx -9\n" +
                "addx 1\n" +
                "addx 1\n" +
                "addx -3\n" +
                "addx 11\n" +
                "noop\n" +
                "noop\n" +
                "addx 1\n" +
                "noop\n" +
                "addx 1\n" +
                "noop\n" +
                "noop\n" +
                "addx -13\n" +
                "addx -19\n" +
                "addx 1\n" +
                "addx 3\n" +
                "addx 26\n" +
                "addx -30\n" +
                "addx 12\n" +
                "addx -1\n" +
                "addx 3\n" +
                "addx 1\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "addx -9\n" +
                "addx 18\n" +
                "addx 1\n" +
                "addx 2\n" +
                "noop\n" +
                "noop\n" +
                "addx 9\n" +
                "noop\n" +
                "noop\n" +
                "noop\n" +
                "addx -1\n" +
                "addx 2\n" +
                "addx -37\n" +
                "addx 1\n" +
                "addx 3\n" +
                "noop\n" +
                "addx 15\n" +
                "addx -21\n" +
                "addx 22\n" +
                "addx -6\n" +
                "addx 1\n" +
                "noop\n" +
                "addx 2\n" +
                "addx 1\n" +
                "noop\n" +
                "addx -10\n" +
                "noop\n" +
                "noop\n" +
                "addx 20\n" +
                "addx 1\n" +
                "addx 2\n" +
                "addx 2\n" +
                "addx -6\n" +
                "addx -11\n" +
                "noop\n" +
                "noop\n" +
                "noop\n";
        smallALU.operation(input.split("\n"));
        assertEquals(21,smallALU.getXByCycleNumber(20));
        assertEquals(19,smallALU.getXByCycleNumber(60));
        assertEquals(18,smallALU.getXByCycleNumber(100));
        assertEquals(21,smallALU.getXByCycleNumber(140));
        assertEquals(16,smallALU.getXByCycleNumber(180));
        assertEquals(18,smallALU.getXByCycleNumber(220));
    }
}